move contrib tests to separate file
